==> notebooks/00_Refine_table.ipynb <==
import pandas as pd

df_new = pd.read_excel("Table_1._Question-SQL_pair_NEW.xlsx", index_col=0)
df_new.label = df_new.label.fillna("Training")
df_new.index.name = "id"
df_new = df_new.reset_index()
df_new.sql = df_new.sql.str.strip().str.replace("\n", " ")
df_new

df_old = pd.read_json("../train/antismash/question_sql_pair.json")

df_new.reset_index()

df_new.to_json("question_sql_pair_benchmark_140824.json", indent=2, orient="records")

df_test = df_new[df_new.label == "Test"]

df_test.to_json("question_sql_pair_test_140824.json", indent=2, orient="records")
