==> notebooks/05_Database_summary.ipynb <==
#! pip install jupysql pandas matplotlib duckdb-engine

import duckdb
import pandas as pd

%load_ext sql
    
conn = duckdb.connect("../data/antismash_db.duckdb")
%sql conn --alias duckdb

# How many genomes
genome = conn.sql("""
SELECT d.accession, g.*, t.* 
FROM antismash.genomes g 
JOIN antismash.taxa t ON g.tax_id = t.tax_id
JOIN antismash.dna_sequences d ON d.genome_id = g.genome_id;
""").df()

genome

table_schema = conn.sql("""
SELECT * 
FROM information_schema.columns 
WHERE table_schema = 'antismash';
"""
).df()
print("table", len(table_schema.table_name.unique()))
print("column", len(table_schema.column_name))

result = []
for t in table_schema.table_name.unique():
    query = f"SELECT COUNT(*) FROM antismash.{t};"
    query = conn.sql(query).fetchone()
    if query[0] > 500000:
        print(t, query)
    result.append(query[0])
print("rows", sum(result))

max(result)

df_bgcflow = pd.read_csv("../data/samples_genbank_with_metadata_edited.csv")
df_bgcflow[~df_bgcflow.genome_id.isin(genome.assembly_id)].genome_id

genome

regions = conn.sql("""
SELECT * 
FROM antismash.regions r
JOIN antismash.rel_regions_types rrt ON r.region_id = rrt.region_id
JOIN antismash.bgc_types bt ON rrt.bgc_type_id = bt.bgc_type_id;
""").df()
regions.category.value_counts()

regions.contig_edge.value_counts(normalize=True) * 100

regions

mibig_hits = regions[regions.best_mibig_hit_similarity > 80]
print(len(mibig_hits), len(mibig_hits) / len(regions)*100)
mibig_hits.best_mibig_hit_description.value_counts()[:10]

mibig_hits = regions[regions.best_mibig_hit_similarity.isnull()]
print(len(mibig_hits), len(mibig_hits) / len(regions)*100)
mibig_hits.best_mibig_hit_description.value_counts()[:10]
