==> notebooks/06_Supplementary_materials.ipynb <==
import pandas as pd

df = pd.read_csv("2.Evaluated_benchmark_response.tsv", sep="\t", index_col=0)

df.columns

df.loc[:, ["question", "difficulty", 'chatbgc_version',
       'benchmark_version', 'sql_query_success', 'summary_evaluation',
       'sql_evaluation', 'evaluator_model', 'summary_evaluation_reason',
       'sql_evaluator_model']]
